Device_Embedding/RPI/Robotics/requirements.txt
RPi.GPIO
Adafruit-Blinka
Adafruit-BNO055
